creation du main2a pour les test des fonctions elementaires. globalement tout est codé. pour l'instant tests juste pour les quadrangles, a verifier pour segments surtout. Quelques corrections dans les fonctions et ajout de nouvelles fonctions pour simplifier les test. Ne pas tenir compte de mon dernier commit j'avais oublier de add le main2a (oups). Reste la doc de quelques fonctions et à mettre bien propre les commentaires etc, je regarde ça dans le we.

+++ elementaires.h
@@ -120,7 +120,7 @@ void freetab(void *ptr);
 
 float *allocvec_f(int dim);
 
-float *allocvec_i(int dim);
+int *allocvec_i(int dim);
 
 void printMat(int m, int n, float **mat);
